ffglitch: libavutil/script: fix running of quickjs scripts from different threads
This bug happened on Windows, macOS, and under Valgrind on Linux (but not
under normal operation under Linux), where the mb_type and pict_type
features would exit with a (non-descriptive) exception, caused by a stack
check overflow.

The reason this happened was because quickjs checks the stack in a way
that's not very thread-friendly, so the stack pointer needs to be updated
every time the same JSRuntime is used in a different thread.

Note that quickjs doesn't support multithreading with the same JSRuntime,
this is entirely single-threaded code, but the current calling thread may
change, which was the case with mb_type and pict_type.